#16 Transferred the verification logic from the KeeperRequest to the Controller
